Frankensteins Monster lebt

+++ Model/AccountEntry.cs
@@ -1,9 +1,9 @@
 ﻿namespace DNPL.Model;
 public struct AccountEntry {
 
-	public double Before { get; set; }
-	public double Amount { get; set; } //Amount
-	public double After { get; set; }
+	public decimal Before { get; set; }
+	public decimal Amount { get; set; } //Amount
+	public decimal After { get; set; }
 	public string Type { get; set; } //BalanceChangeType
 	public DateTime Date { get; set; }
